fixed some minor bugs; each MapGenerator implementation can now specify a default start position and zoom level; improved the performance of the Cohen-Sutherland line clipping implementation; added Javadoc comments to some classes and methods; deleted unused Projection interface.
git-svn-id: http://mapsforge.googlecode.com/svn/trunk@151 a424f762-2dd8-11df-bfeb-7b4c664f286a

+++ mapsforge/src/org/mapsforge/android/map/ImageFileCache.java
@@ -90,10 +90,14 @@ class ImageFileCache {
 	 */
 	synchronized void destroy() {
 		// delete all cached files
-		for (File file : this.map.values()) {
-			if (!file.delete()) {
-				file.deleteOnExit();
+		if (this.map != null) {
+			for (File file : this.map.values()) {
+				if (!file.delete()) {
+					file.deleteOnExit();
+				}
 			}
+			this.map.clear();
+			this.map = null;
 		}
 	}
 
@@ -137,10 +141,10 @@ class ImageFileCache {
 	 */
 	synchronized void setCapacity(int capacity) {
 		this.capacity = capacity;
-		// make a new map with the new capacity and put all entries from the old map in the new
-		// one. The replacement policy will keep the right files in the new map if the new map
-		// has a smaller capacity than the old map.
+		// create a new map with the new capacity
 		LinkedHashMap<Tile, File> newMap = createMap(this.capacity);
+
+		// put all entries from the old map in the new one.
 		for (Map.Entry<Tile, File> entry : this.map.entrySet()) {
 			newMap.put(entry.getKey(), entry.getValue());
 		}
